% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_daily_sequence}
\alias{get_daily_sequence}
\title{Return a Sequence of Days}
\usage{
get_daily_sequence(start_date, end_date)
}
\arguments{
\item{start_date}{Starting date. Required}

\item{end_date}{End date. Required}
}
\value{
A vector of class 'Date'
}
\description{
Return a Sequence of Days
}
\seealso{
\code{\link[=seq.Date]{seq.Date()}} which this function wraps.
}
