% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_load_profile.R
\name{get_load_profile}
\alias{get_load_profile}
\title{Generate a load profil}
\usage{
get_load_profile(
  profile = c("H0", "G0", "G1", "G2", "G3", "G4", "G5", "G6", "L0", "L1", "L2"),
  start_date = Sys.Date(),
  end_date = Sys.Date() + 2
)
}
\arguments{
\item{profile}{Name of the load profile, see 'Details'.}

\item{start_date}{starting date in ISO 8601 date format, required}

\item{end_date}{end date in ISO 8601 date format, required}
}
\value{
A data.frame with two columns:
\itemize{
\item 'date', class POSIXct
\item 'load_profile', in Watt
}
}
\description{
Generate a load profil
}
\details{
A load profile describes the consumption profile of an electricity
consumer over a certain period of time. This function creates a data sequence
based on the user input and generates a load profile for the requested
'profile'.

Supported are:
\itemize{
\item H0: households (German: "Haushalte")
\item G0 to G6: commerce ("Gewerbe")
\item L0 to L2: agriculture ("Landwirtschaft")
}

Call \code{\link[=get_load_profile_info]{get_load_profile_info()}} for more information on each profile and examples.
}
\examples{
get_load_profile()
get_load_profile("G0", "2024-01-01", "2024-01-31")

}
