% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profile_info.R
\name{get_profile_info}
\alias{get_profile_info}
\title{Get the description load profiles according to the BDEW}
\source{
\url{https://www.bdew.de/energie/standardlastprofile-strom/}
}
\usage{
get_profile_info(language = c("EN", "DE"))
}
\arguments{
\item{language}{One of 'EN' (English), 'DE' (German)}
}
\value{
A data.frame
}
\description{
Get the description load profiles according to the BDEW
}
\examples{
get_profile_info(language = "EN")
get_profile_info(language = "DE")
}
