% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profile_information.R
\name{get_profile_information}
\alias{get_profile_information}
\title{Get the description of 11 standarized load profiles according to the BDEW}
\source{
\url{https://www.bdew.de/energie/standardlastprofile-strom/}
}
\usage{
get_profile_information(language = c("english", "german"))
}
\arguments{
\item{language}{One of 'english', 'german'}
}
\value{
A data.frame
}
\description{
Get the description of 11 standarized load profiles according to the BDEW
}
\examples{
get_profile_information(language = "english")
get_profile_information(language = "german")
}
