% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weekday.R
\name{get_weekday}
\alias{get_weekday}
\title{Map a vector of class 'Date' to day of the week to each value in 'x'}
\usage{
get_weekday(x)
}
\arguments{
\item{x}{A sequence of class 'Date'}
}
\value{
A character vector
}
\description{
Map a vector of class 'Date' to day of the week to each value in 'x'
}
