% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{slp}
\alias{slp}
\title{The \code{slp} dataset was renamed}
\description{
\strong{Defunct.} The electricity dataset \code{slp} was renamed to \link{slp_electricity_profiles} in
version 2.0.0. Accessing \code{slp} now raises an error. Use
\code{slp_electricity_profiles}, or \code{standardlastprofile::slp_electricity_profiles},
instead.
}
\seealso{
\link{slp_electricity_profiles}
}
\keyword{internal}
