% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardlastprofil-package.R
\docType{package}
\name{standardlastprofil-package}
\alias{standardlastprofil}
\alias{standardlastprofil-package}
\title{standardlastprofil: Generate load profiles for electricity}
\description{
For electricity customers without advanced metering infrastructure, an energy supplier uses standardised load profiles (SLP). These SLPs replace the non-existent load profile of end consumers with a calculated, sufficiently accurate forecast of electricity consumption on a 15-minutes basis. This packages provides the the German Association of Energy and Water Industries (BDEW Bundesverband der Energie- und Wasserwirtschaft e.V.) in a tidy format. The data contains standardized load profiles for three customer groups -- household, agriculture and commercial customers. A distinction is also made between weekdays, Saturdays and Sundays and between summer, winter and a transitional period.
}
\author{
\strong{Maintainer}: Markus Döring \email{m4rkus.doering@gmail.com}

}
\keyword{internal}
