% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{german_states}
\alias{german_states}
\title{German state names}
\format{
A data.frame with 3 variables: \code{state_code}, \code{state_de}, and \code{state_en}
\describe{
\item{state_code}{character, ISO 3166-2:DE code for all German states}
\item{state_de}{state name in German}
\item{state_en}{state name in English}
}
}
\source{
\url{https://www.iso.org/iso-3166-country-codes.html}

\url{https://en.wikipedia.org/wiki/States_of_Germany}
}
\usage{
german_states
}
\description{
The Federal Republic of Germany, as a federal state, consists of 16 states.
This dataset returns all 16 state codes in ISO 3166 standard and their respective
name in German and English.
}
\examples{
german_states

}
\keyword{datasets}
