% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_sequence.R
\name{get_daily_sequence}
\alias{get_daily_sequence}
\title{Generate a daily time series}
\usage{
get_daily_sequence(start_date, end_date)
}
\arguments{
\item{start_date}{Starting date. Required}

\item{end_date}{End date. Required}
}
\value{
A vector of class "Date".
}
\description{
Generate a daily time series
}
\examples{
today <- Sys.Date()
get_daily_sequence(today, today + 3)
}
