% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_load_profile.R
\name{get_load_profile}
\alias{get_load_profile}
\title{Generate a Load Profile}
\source{
\url{https://www.bdew.de/energie/standardlastprofile-strom/}

\url{https://www.bdew.de/media/documents/1999_Repraesentative-VDEW-Lastprofile.pdf}

\url{https://www.bdew.de/media/documents/2000131_Anwendung-repraesentativen_Lastprofile-Step-by-step.pdf}
}
\usage{
get_load_profile(
  profiles = c("H0", "G0", "G1", "G2", "G3", "G4", "G5", "G6", "L0", "L1", "L2"),
  start_date,
  end_date
)
}
\arguments{
\item{profiles}{name of load profiles, see 'Details'.}

\item{start_date}{start date in ISO 8601 date format, required}

\item{end_date}{end date in ISO 8601 date format, required}
}
\value{
A data.frame with four variables:
\itemize{
\item \code{profile}, character, load profile identifier
\item \code{start_time}, POSIXct / POSIXlt, start time
\item \code{end_time}, POSIXct / POSIXlt, end time
\item \code{watts}, numeric, electric power
}
}
\description{
\code{get_load_profile()} returns load profiles in the form of quarter-hourly
values (in Watt) that are standardized to an annual consumption of 1,000 kWh.
}
\details{
Supported profiles are:
\itemize{
\item \code{H0}: households (German: "Haushalte")
\item \code{G0} to \code{G6}: commerce ("Gewerbe")
\item \code{L0} to \code{L2}: agriculture ("Landwirtschaft")
}

Call \code{\link[=get_load_profile_info]{get_load_profile_info()}} for more information about profiles.

The standard load profiles are differentiated according to winter,
transitional period, summer as well as workday, Saturday, Sunday. See:
\url{https://www.bdew.de/media/documents/1999_Repraesentative-VDEW-Lastprofile.pdf}
for the methodology used to determine the profiles.

Periods are defined as:
\itemize{
\item \code{summer}: May 15 to September 14
\item \code{winter}: November 1 to March 20
\item \code{transition}: March 21 to May 14, and September 15 to October 31
}

Dates are mapped to one of three days:
\itemize{
\item \code{saturday}: Saturdays; Dec 24th and Dec 31th are considered a Saturday too,
if they are not a Sunday
\item \code{sunday}: Sundays and all public holidays
\item \code{workday}: Monday to Friday
}

\strong{Note}: The package supports nationwide, public holidays for Germany only. Those
were retrieved from the \href{https://github.com/nager/Nager.Date}{nager.Date API} and
are listed below for 2024:
\itemize{
\item Jan 1: New Year's
\item Mar 29: Good Friday
\item Apr 1: Easter Monday
\item May 1: Labour Day
\item May 9: Ascension Day
\item May 20: Whit Monday
\item Oct 3: German Unity Day
\item Dec 25: Christmas Day:
\item Dec 26: Boxing Day
}
}
\examples{
today <- Sys.Date()
get_load_profile("H0", today, today + 1) |> head()
L <- get_load_profile(c("L0", "L1", "L2"), today, today + 1)
head(L)
}
