% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_period}
\alias{get_period}
\title{Map a sequence of dates to a consumption period according to BDEW definition}
\source{
description
}
\usage{
get_period(x)
}
\arguments{
\item{x}{A vector of class 'Date'}
}
\value{
A character vector; a vector of dates mapped to 'summer', 'winter', or 'transition'
}
\description{
Map a sequence of dates to a consumption period according to BDEW definition
}
