% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weekday.R
\name{get_weekday}
\alias{get_weekday}
\title{Map a vector of class 'Date' to 'working_day', 'saturday', or 'sunday'}
\usage{
get_weekday(x)
}
\arguments{
\item{x}{A sequence of class 'Date'}
}
\value{
A character vector; a vector of dates mapped to 'working_day', 'saturday', or 'sunday'
}
\description{
Map a vector of class 'Date' to 'working_day', 'saturday', or 'sunday'
}
