% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{load_profiles}
\alias{load_profiles}
\title{Load Profile Data for Electricity from BDEW}
\format{
A data frame with 9,504 rows and 5 columns:
\describe{
\item{profile}{character, one of 11 load profiles, see 'Details'}
\item{period}{factor, one of 'summer', 'winter', 'transition'}
\item{day}{factor, one of 'saturday', 'sunday', and 'workday'}
\item{timestamp}{character, format \%H:\%M}
\item{watt}{numeric, electricity load}
}
}
\source{
\url{https://www.bdew.de/energie/standardlastprofile-strom/}

\url{https://www.bdew.de/media/documents/1999_Repraesentative-VDEW-Lastprofile.pdf}
}
\usage{
load_profiles
}
\description{
A load profile shows the variation in the electrical load versus time. Data is
provided by the German Association of Energy andvWater Industries (BDEW
Bundesverband der Energie- und Wasserwirtschaft e.V.).
}
\details{
The load profiles are available in the form of quarter-hourly values in Watt
that are standardised to an annual consumption of 1,000 kWh. They are
differentiated by workday, Saturday and Sunday in the three annual periods of
winter, transitional period and summer. Hence for each profile there are 96
values per weekday and period. See 'Details' for more information.

Included are eleven representative load profiles covering three different
customer groups:
\itemize{
\item households: \code{H0}
\item commercial: \code{G0}, \code{G1}, \code{G2}, \code{G3}, \code{G4}, \code{G5}, \code{G6}
\item agriculture: \code{L0}, \code{L1}, \code{L2}
}

Call \code{\link[=get_load_profile_info]{get_load_profile_info()}} to get description for each profile.

Periods as defined by the BDEW:
\itemize{
\item summer: May 15 to September 14
\item winter: November 1 to March 20
\item transition: March 21 to May 14, and September 15 to October 31
}

In addition, a distinction is made between three different type of days:
\itemize{
\item \code{workday}: Monday to Friday
\item \code{saturday}: Saturdays; Dec 24th and Dec 31th are considered a Saturday too,
if they are not a Sunday
\item \code{sunday}'`: Sundays and all nationwide holidays in Germany
}
}
\keyword{datasets}
