% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{load_profiles}
\alias{load_profiles}
\title{Data of load profiles for electricity}
\format{
\subsection{\code{load_profiles}}{

A data frame with 9,504 rows and 5 columns:
\describe{
\item{profile}{one of 11 load profiles, see 'Details'}
\item{period}{one of 'summer', 'winter', 'transition'}
\item{weekyday}{one of 'saturday', 'sunday', and 'working_day'}
\item{timestamp}{timestemp in 15-minutes interval}
\item{value}{electricity load in Watt.}
}
}
}
\source{
\url{https://www.bdew.de/energie/standardlastprofile-strom/}

\url{https://www.bdew.de/media/documents/1999_Repraesentative-VDEW-Lastprofile.pdf}
}
\usage{
load_profiles
}
\description{
A load profile shows the variation in the electrical load versus time. Data is
provided by the German Association of Energy andvWater Industries (BDEW
Bundesverband der Energie- und Wasserwirtschaft e.V.). A load profile
shows the variation in the electrical load versus time. For each profile
there are 96 values for three different days of the week ('working_day',
'saturday', 'sunday'), three different periods of the year ('summer', 'winter',
'transition'), i.e. a total of 3x3x96 observations per profile. See 'Details'
for more information.
}
\details{
Included are eleven representative load profiles covering three different
customer groups:
\itemize{
\item households: 'H0'
\item commercial: 'G0', 'G1', 'G2', 'G3', 'G4', 'G5', 'G6'
\item agriculture: 'L0', 'L1', 'L2'
}

Call \code{\link[=get_load_profile_info]{get_load_profile_info()}} to get description for each profile.

Periods as defined by the BDEW:
\itemize{
\item summer: May 15 to September 14
\item winter: November 1 to March 20
\item transition: March 21 to May 14, and September 15 to October 31
}

In addition, a distinction is made between three different type of days:
\itemize{
\item 'working_day': Monday to Friday
\item 'saturday': Saturdays; Dec 24th and Dec 31th are considered a 'saturday' too, if they are not a 'sunday'
\item 'sunday': Sundays and all public holidays
}
}
\keyword{datasets}
