% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{slp}
\alias{slp}
\title{The \code{slp} dataset was renamed}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
The electricity dataset \code{slp} was renamed to \link{slp_electricity_profiles} in
version 2.0.0 and is no longer exported under the old name. Accessing \code{slp}
still returns the data for now, but emits a deprecation warning. Use
\code{slp_electricity_profiles}, or \code{standardlastprofile::slp_electricity_profiles},
instead.
}
\seealso{
\link{slp_electricity_profiles}
}
\keyword{internal}
