% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slp_info.R
\name{slp_info}
\alias{slp_info}
\title{Information and Examples of BDEW Standard Load Profiles}
\source{
\url{https://www.bdew.de/energie/standardlastprofile-strom/}

\url{https://www.bdew.de/media/documents/2000131_Anwendung-repraesentativen_Lastprofile-Step-by-step.pdf}

\url{https://www.bdew.de/media/documents/Zuordnung_der_VDEW-Lastprofile_zum_Kundengruppenschluessel.pdf}
}
\usage{
slp_info(language = c("EN", "DE"))
}
\arguments{
\item{language}{one of 'EN' (English), 'DE' (German)}
}
\value{
A data.frame
}
\description{
Information and Examples of BDEW Standard Load Profiles
}
\examples{
slp_info(language = "EN")$G6
slp_info(language = "DE")$G6

}
