% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slp_info.R
\name{slp_info}
\alias{slp_info}
\title{Get the description load profiles according to the BDEW}
\source{
\url{https://www.bdew.de/energie/standardlastprofile-strom/}

\url{https://www.bdew.de/media/documents/2000131_Anwendung-repraesentativen_Lastprofile-Step-by-step.pdf}
}
\usage{
slp_info(language = c("EN", "DE"))
}
\arguments{
\item{language}{one of 'EN' (English), 'DE' (German)}
}
\value{
A data.frame
}
\description{
Get the description load profiles according to the BDEW
}
\examples{
slp_info(language = "EN")$G6
slp_info(language = "DE")$G6

}
