% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardlastprofil-package.R
\docType{package}
\name{standardlastprofile-package}
\alias{standardlastprofile}
\alias{standardlastprofile-package}
\title{standardlastprofile: R Data Package for Load Profiles for Electricity}
\description{
Data on standard load profiles from the German Association of Energy and Water Industries (BDEW Bundesverband der Energie- und Wasserwirtschaft e.V.) in a tidy format. The package also provides an interface to generate a load profile over a user specified period of time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flrd/standardlastprofile}
  \item \url{https://flrd.github.io/standardlastprofile/}
  \item Report bugs at \url{https://github.com/flrd/standardlastprofile/issues}
}

}
\author{
\strong{Maintainer}: Markus Döring \email{m4rkus.doering@gmail.com} [copyright holder]

}
\keyword{internal}
