% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardlastprofil-package.R
\docType{package}
\name{standardlastprofile-package}
\alias{standardlastprofile}
\alias{standardlastprofile-package}
\title{standardlastprofile: R data package for load profiles for electricity}
\description{
This package provides data on standardized load profiles from the German Association of Energy and Water Industries (BDEW Bundesverband der Energie- und Wasserwirtschaft e.V.) in a tidy format. The data was published in 1999 and since then has formed a basis for energy suppliers to approximate the estimated electricity consumption for their customers. A standardized load profile describes the consumption profile of an energy consumer over a certain period of time. The package also provides an interface to generate a load profile over a user specified period of time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flrd/standardlastprofile}
  \item \url{https://flrd.github.io/standardlastprofile/}
  \item Report bugs at \url{https://github.com/flrd/standardlastprofile/issues}
}

}
\author{
\strong{Maintainer}: Markus Döring \email{m4rkus.doering@gmail.com}

}
\keyword{internal}
