% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardlastprofil-package.R
\docType{package}
\name{standardlastprofile-package}
\alias{standardlastprofile}
\alias{standardlastprofile-package}
\title{standardlastprofile: Data Package for BDEW Standard Load Profiles in Electricity}
\description{
Data on standard load profiles from the German Association of Energy and Water Industries (BDEW Bundesverband der Energie- und Wasserwirtschaft e.V.) in a tidy format. The data and methodology are described in VDEW (1999), "Repräsentative VDEW-Lastprofile", \url{https://www.bdew.de/media/documents/1999_Repraesentative-VDEW-Lastprofile.pdf}. The package also offers an interface for generating a standard load profile over a user-defined period. For the algorithm, see VDEW (2000), "Anwendung der Repräsentativen VDEW-Lastprofile step-by-step", \url{https://www.bdew.de/media/documents/2000131_Anwendung-repraesentativen_Lastprofile-Step-by-step.pdf}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flrd/standardlastprofile}
  \item \url{https://flrd.github.io/standardlastprofile/}
  \item Report bugs at \url{https://github.com/flrd/standardlastprofile/issues}
}

}
\author{
\strong{Maintainer}: Markus Döring \email{m4rkus.doering@gmail.com} [copyright holder]

}
\keyword{internal}
